refactor(analysis): split content complexity cognitive tests
